pyproject.toml
[project]
name = "lorchestra"
version = "0.1.0"
description = "lorchestra: Local Orchestrator for PHI data pipeline (meltano → canonizer → vector-projector)"
authors = [{name = "Local Pipeline Team"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "ruff>=0.1",
]

[project.scripts]
lorchestra = "lorchestra.cli:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["lorchestra*"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"
